python/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyzmq>=25.0.0
msgpack>=1.0.5
uvloop>=0.19.0
polars>=0.19.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0

